pyproject.toml
[project]
name = "project-1-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.119.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.37.0",
]
